cvdet: lfacesyn and retiming done

+++ run.py
@@ -45,8 +45,8 @@ def step1_lipsyn(pass_id, train, predict, inp_id, outp_norm=False, preprocess=Fa
     apath   = '%s/mfcc/a%s.npy' % (inp_dir, inp_id)               # audio mfcc
     mpath   = '%s/mp3/a%s.mp3'  % (inp_dir, inp_id)               # music mp3
     opath   = '%s/ldmk/%s/a%s.npy' % (inp_dir, pass_id, inp_id)   # output npy data
-    tmppath = '%s/vis/%s/a%s.avi' % (inp_dir, pass_id, inp_id)    # temporary visualized mp4
-    vpath   = '%s/vis/%s/a%s.mp4' % (inp_dir, pass_id, inp_id)    # final visualization
+    tmppath = '%s/visual/%s/a%s.avi' % (inp_dir, pass_id, inp_id) # temporary visualized mp4
+    vpath   = '%s/visual/%s/a%s.mp4' % (inp_dir, pass_id, inp_id) # final visualization
 
     a2v_cvter = lipnn.Audio2Video(args=args)
     if train:
@@ -94,21 +94,22 @@ def step4_composite(inp_id, tar_id, sq, padw=100, rt=True):
 
 def clean_tmp(pass_id, inp_id, tar_id):
     import os, glob
-    f1 = glob.glob('%s/vis/%s/*.avi' % (inp_dir, pass_id))
+    f1 = glob.glob('%s/visual/%s/*.avi' % (inp_dir, pass_id))
     f2 = glob.glob('%s/a%st%s/*.avi' % (outp_dir, inp_id, tar_id))
     for f in (f1 + f2):
-        os.remove(os.path.abspath(f))
-        print('Remove: %s' % os.path.abspath(f))
+        if os.path.exists(f):
+            os.remove(os.path.abspath(f))
+            print('Remove: %s' % os.path.abspath(f))
     print('Done')
 
 if __name__ == '__main__':
     pass_id = "std_u"
     inp_id  = "015"
     tar_id  = "187"
-    sq = a2v.Square(0.2, 0.8, 0.4, 1.15)
+    sq = a2v.Square(0.3, 0.7, 0.5, 1.1)
     # step0_dataset()
     # step1_lipsyn(pass_id=pass_id, train=False, predict=True, inp_id=inp_id, outp_norm=False)
-    step2_lfacesyn(pass_id, inp_id, tar_id, sq)
+    # step2_lfacesyn(pass_id, inp_id, tar_id, sq)
     # step3_retiming(inp_id, tar_id)
-    # step4_composite(inp_id, tar_id, sq)
+    step4_composite(inp_id, tar_id, sq)
     # clean_tmp(pass_id, inp_id, tar_id)
